Add Ability to Visualize Partial LeRobot Dataset (#11228)
### Related
This is a subset of https://github.com/rerun-io/rerun/pull/11226 where I
am not including the MP4 changes that need more discussion.

### What
Before when I tried to open a partial LeRobot dataset (if I only had a
subset of the contents listed in the metadata) we got all kinds of weird
errors.

I made the episode collection a map instead of a vector which allows us
to display an arbitrary subset of episodes from a dataset.