Fix `to_arrow_reader` causing an error for streaming providers (#11267)
### Related

The `DataFrameQueryView::to_arrow_reader` was broken after switching to
a non-collecting table provider. This was discovered during release
testing.

### What

This PR fixes two issues:

- `PartitionStreamExec` could not be copied using `with_children`
because it would error even if the children vec was empty.
- `to_arrow_reader` was using a blocking call on the python tokio
runtime, but we cannot spawn our additional streams in this way. As an
overall improvement we can instead use the fact that DataFusion
DataFrames already support the arrow C interface for streaming. Thus we
can use `pyarrow.RecordBatchReader.from_stream` to turn a dataframe
directly into a streaming provider. This also has the advantage of
removing the collect operation, so it is not memory bound.