Use `FetchChunks` instead of `GetChunks` for chunk fetching on query execution path (#11215)
### What

We've recently added support for the new ``FetchChunks`` API on the
Rerun Cloud side that should be more optimal and avoid re-querying
internal indexes. As part of initial ``QueryDataset`` request we're
asking the server to give us ``ChunkKey`` information that holds
location details we'll send back to the server as part of
``FetchChunks``.

``FetchChunks`` is now leveraged instead of ``GetChunks`` and it uses
relevant chunk information (``ChunkKey``) to fetch the actual data from the
backend. Rerun Cloud has simpler and more optimal logic for fetching chunk data
through ``FetchChunks`` API hence this comes with performance gains and better
future scaling.