Add support for `enum` in protobuf MCAP messages (#11280)
### Related

* Related #11221. Did not end up using `UnionBuilder` in this PR, but
will be required for `oneof`.
* [x] Requires #11278.
* [x] Requires #11283.

### What

This adds `enum` support to the decoding of protobuf MCAP messages. It
does not make use of `UnionBuilder` as originally planned, because that
would be more inefficient then using plain `StringBuilder` (we can't
share union variants, or Arrow dictionaries across lists in our data
model).

This also replaces our asset-based protobuf snapshot test with a basic
in-memory test harness. This will allow iterating on the tests in the
future more quickly.