Remove `ItemButton`s on favor of new `OnResponse` helper (#11233)
### Related

* follow up of https://github.com/rerun-io/rerun/pull/11164

### What

Adds a new `OnResponseExt` trait that allows you to call `on_click`,
`on_menu`, `enabled` and much more on anything that is `impl Widget`.

We want to experiment with this in rerun and eventually move this to
egui (if we decide that this is a good pattern).
One blocker for moving to egui would be that we can't customize the
default menu_styling right now.