Add table support to OSS server (#11356)
This PR adds support for lance tables in the OSS server, similar to the
existing support for Rerun datasets.

Users can add a `--table file_path` option to provide a path to the
table location.

Additionally it creates the `__entries` table that is required for using
the dataframe APIs.

Unit tests were updated and verified to match the redap server.

This is an enabling feature for ongoing work such as catalog providers
where we need to use the OSS server in CI.