table provider: propagate stream error from IO loop (#11397)
### What

Right now if there's an error in the chunk fetching stream, it will be
silently ignored. Instead here we make a change to propagate results
from the IO loop and handle them in the CPU loop as we already handle
errors in CPU loop. This ensures that receiving an error will actually
result in stream termination (with an error).