Add `-` formatting option for redacted values (#11383)
### Related

* Follow-up of #11278.

### What

This PR allows you to use the minus sign (`-`) formatting option in most
`RecordBatch`-derived structs to get a string that has non-deterministic
values redacted.

Usage:

```rs
let chunk = Chunk::new(/* ... */);
println!("{chunk:-}");
```

Using `#` instead of `-` would be an alternative, but I think it's
better to reserve this for a more meaningful representation.