Add meaningful equality comparisons to many rust wrappers (#11401)
### What
pyo3 defaults `__eq__` to be equivalent to `is` but since we clone
before returning most python objects that means
`dataset.schema()==dataset.schema()` is False even though the contents
are identical.

I resolved this specific case and added a test.
Then I asked a robot to write me a lint to enforce this everywhere,
updated for things I could easily `derive(PartialEq, Eq)` then added
NOLINT to the rest.
The score is
| With eq | Without eq |
|---------|------------|
|      8       |      15           |