Merge `DroppableEncoder` and `Encoder` into a single facility (#11446)
Our RRD encoding facilities are currently implemented by two distinct
types: `DroppableEncoder` & `Encoder`.

This makes understanding how to use the API unnecessarily painful. Turns
out the only reason things are done this way is because of an
implementation detail related to the `Drop` implementation.
Instead of passing off that complexity to end-users of the API, we now
contain it to the implementation instead.

In addition to that, the `AlreadyFinished` logic was broken, so fix too
while we're at it.

---

This PR is part of an upcoming series of PRs to pay off organic growth
debt in our encoding/decoding stack.