`StreamDecoder`: lift naming confusion regarding chunks (#11452)
This is just a pure renaming pass.

`StreamDecoder`'s job is to deal with chunks of bytes, and it used to
call those `chunks`. This was extremely confusing to me when trying to
understand the code, since _chunk_ is such an overloaded term for us.
This PR just makes it as clear as possible that this is about _byte
chunks_, not _Rerun chunks_.

---

This PR is part of an upcoming series of PRs to pay off organic growth
debt in our encoding/decoding stack.

* DNM: requires #11450