`StreamDecoder`: fix support for out-of-order legacy store ID (#11455)
This PR ports the test suite from the legacy `Decoder` to the new
`StreamDecoder`, to ensure they are fully compatible.

In doing so, I stumbled upon a bug where the `StreamDecoder` would enter
an obscure irrecoverable state when faced with an out-of-order legacy
store ID (i.e. a store ID missing an app ID).
This PR fixes that.

Also removed some duplicated code on the way.

---

This PR is part of an upcoming series of PRs to pay off organic growth
debt in our encoding/decoding stack.

* DNM: requires #11452