Introduce `Lens` as a concept for semantic mapping of arbitrary data (#11394)
### Related

* [x] Requires #11383.
* [x] Requires #11456.
* [x] Requires #11454.

### What

This PR introduces the concept of _Lenses_ as a way to express
re-interpretation and manipulation of component columns. This can, for
example, be useful to attach Rerun semantics to arbitrary Arrow
components that are created when loading MCAP files.

For now, the only way to access the feature is via the `LensesSink`,
which transforms component columns and then forwards aggregated chunks
to the backing sink.

> [!IMPORTANT]
> Lenses are still a very experimental feature and we expect the API to
change a lot in upcoming releases.