Make time and timeline controlled by blueprint (#11405)
### Related

* Closes #7728

### What

Makes time cursor and current timeline be controlled by the blueprint.

### Details

We log the time as static for blueprints for it not to affect undo
state.

This includes some refactors that felt necessary to make tests still
work and not be overly complicated:
- `TestContext` doesn't require a mutable reference to run.
- Can update `TestContext` blueprints via
`TestContext::with_blueprint_ctx`.
- The blueprint ctx is a struct that implements a helper trait for
interacting with blueprints. App also has an equivalent
`AppBlueprintCtx`, and `ViewerContext` also implements this trait.
- `TimeControl` is no longer stored behind an `RwLock` in app, and
instead interacted with through sending
`SystemCommand::TimeControlCommands`.

The time panel was refactored a bit to allow the normal time panel and
blueprint inspector time panel to still share code. This is mostly
changing `&mut TimeControl` to `&mut impl TimeControlCtx` which handles
writing to blueprints/just updating time control whether this is the
blueprint inspector time panel or not.

### TODO
- [x] Fix times per timeline with blueprint stored time, probably not
store time per timeline in blueprint now and instead "switch out" what
time to use.
- [x] Fix snapshot tests, the current produced results are most likely
wrong.

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>