Introduce all-in-one `StreamDecoder` iterator and use it _everywhere_ (#11464)
This PRs augments the existing `StreamDecoder` with the couple features
that it was still missing, and introduce a new `StreamDecoderIterator`
that builds on top of it to provide an all-in-one fully-featured decoder
that supports all use cases.

All in all, this new all-in-one decoder supports (in no particular
order):
* Native and web, both with or without control of the underlying IO loop
* Tail mode, using end-of-stream markers (when coupled with a
`RetryableFileReader`)
* Support for concatenated file streams, with and without EOS markers
(!)
* Protection against under and over IO buffering by requiring `BufRead`
instead of `Read`
* Capable of generically switching between transport and application
level decoding
* Generic support for all the weird things: version propagation, app ID
patching, etc
* Eager initialization for early error detection

Whether we should or shouldn't support all these features in the first
place is a topic for another day (specifically: when we start looking
into RRD footers).
For now, the point is each of these features are required by one or more
entrypoints in the existing OSS codebase, and therefore the goal for now
is to reach parity so have one decoder to rule them all.

The focus here is purely on OSS: for now, the dataplatform decoder
(`StreamingDecoder`) is still completely independent. I'm hoping I can
apply the same kind of reconciliation on that end in a second pass.

--- 

This PR is part of an upcoming series of PRs to pay off organic growth
debt in our encoding/decoding stack.

* DNM: requires #11455
* Closes #11266

---

* [x] `@rerun-bot full-check`
* [x] `cat ./tests/assets/rrd/examples/* | pixi run rerun rrd merge |
pixi run rerun -`
* [x] `cat ./tests/assets/rrd/examples/* | pixi run rerun rrd route
--application-id my_app --recording-id my_rec | pixi run rerun -`