Implement shallow RRD pattern definition for ImHex (#11506)
This is an incomplete Pattern Language definition for the
[ImHex](https://imhex.werwolv.net/) editor. It came to life over the
week-end as I was having night terrors of RRD corruption, and had been
looking for an excuse to learn ImHex for a while.

To use it, load any RRD file in the ImHex editor, and import this file
as a pattern definition.

Unfortunately this merely covers the very outer layers of the Rerun RRD
file encoding.
The RRD protocol relies on several layers of encodings that require code
generation as well as third party libraries. To provide a proper pattern
definition for the RRD protocol, not only a full Rerun ImHex plugin
would be required (and that's C++), but the existing codebase would also
need custom Arrow and Protobuf decoders to map byte spans to actual
column names, etc.
Still, it's a start. May it serves as breadcrumbs for the next soul
venturing into this.

Here's an example of a `dna.rrd`:
<img width="3841" height="2161" alt="image"
src="https://github.com/user-attachments/assets/c171aa78-cb3a-4536-98ae-1825acbc74b8"
/>