Add kitdiff comment to PRs (#11282)
This adds a comment to the web-viewer pr comment action that links to
the kitdiff viewer to make it easy to review snapshot changes. Kitdiff
needs some more work though.

There also are some changes to theme.ron to have some diff to test with,
which should be removed before merging.

- closes https://github.com/rerun-io/rerun/issues/10991