Add update snapshots workflow (#11549)
### Related

- part of
https://linear.app/rerun/issue/RR-1611/make-it-easier-to-update-kittest-images

### What

This adds two scripts: 
- `accept_snapshots`: Move all `.new.png` files to `.png`, updating the
kittest snapshots
- `update_snapshots_from_ci`: Search for the last ci run that matches
the branch name, download the ubuntu test results artifact, unzip it, so
that we have the .new.png files. Then run `accept_snapshots`.

Adds a new workflow that calls `update_snapshots_from_ci` with the
provided RUN_ID. This can be called via a button in kitdiff.