add table writing to lance table providers (#11565)
### Related

Addresses part of
https://linear.app/rerun/project/write-tables-to-rerun-cloud-via-the-python-sdk-432c0243348f/overview

### What

This PR is foundational to our ability to write to any kind of table on
the data platform. Included in this PR is enabling `insert_into` for our
gRPC based table providers. This also includes updates to the OSS server
to support table writing.

We cannot merge this PR until the following are completed:

- [x] Add identical support to data platform
- [x] Add python integration tests