py-sdk: add register_prefix SDK call (#11600)
### Related

* Closes RR-1962
* Follows:
  * https://github.com/rerun-io/rerun/pull/11564
  * https://github.com/rerun-io/dataplatform/pull/1864

### What

This PR builds up on the previous work and adds a new `register_prefix`
python
SDK function that allows end-users to register all the recordings under
a given
object store prefix (i.e., a directory).

### Why

In a previous version of this change, we allowed users to register a
prefix
simply by implicitly interpreting any trailing '/' in the object store
URL as meaning
the customer wanted to register a prefix.

As debated in #11564, however, this is error prone, so we moved to
requiring an explicit signal.

Instead of modifying the signature of the existing `register_batch`
function in non-obvious
and possibly odd ways, I decided to add a new function that is explicit
in its intent, that is.
`register_prefix`.


### FAQs
* The redap API allows prefix and non-prefix `DataSources` to be mixed
in the same call. Why does
  the python SDK not allow this?
* We'd need to make the `register_batch` python function signature more
complex and awkward.
This keeps the user surface simple, while leaving room for future
improvements if needed.
* Why not `register_prefixes` (plural)?
* We could, but I don't see a strong use case for registering multiple
prefixes
     at once.


### Testing
Manually

---------

Signed-off-by: Andrea Reale <andrea@rerun.io>