When scrubbing time, intelligently round to a nice value (#11658)
### Related
* Part of
https://linear.app/rerun/issue/RR-2768/improve-time-selection-experience-on-a-timeline

### What
When dragging the time cursor ("time scrubbing") we would previously
pick arbitrarily precise time points, regardless of zoom level. Now, we
snap to the "simplest" value that is within a pixel from the mouse
cursor. The simplest value is defined as the one with _the most_
trailing zeroes (e.g. prefer `23.500000` over `23.484793`).

### Before
Even though we zoomed out to where we are viewing full hours, when we
try to pick a time we select with microsecond precision:


![dumb-aim](https://github.com/user-attachments/assets/22a10523-a7a9-4474-909f-e61e11a8f551)


### After
Notice how we start by picking whole multiples of 10s, but as we zoom in
we pick multiples of 0.5s, and then 5ms, 10us, etc.


![smart-aim](https://github.com/user-attachments/assets/d9508f33-b724-43fe-8037-a4f5f33c0480)

Note that we only ever omit _zeroes_ in the display, so this is a real
difference in how values are picked, NOT how they are displayed!


### Future work
* Coarser snapping when a modifier key is pressed
* Set the time display precision based on zoom level