Add a new `target_partition_num_rows` API parameter to the CreateIndex API  (#11686)
### What

The old way of specifying `num_partitions` is now deprecated.
See:
https://docs.rs/lance-index/0.38.2/lance_index/vector/ivf/builder/struct.IvfBuildParams.html#structfield.num_partitions

Lance exposes a new `target_partition_size` parameter which gives me
predictable performance instead of users manually needing to increase
the number of partitions as the size of the dataset increases. This
works better for things like incremental indexing.

Resolves: RR-2772

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>