Create lance tables in the cloud (#11640)
### Related

This is a follow on to https://github.com/rerun-io/rerun/pull/11565

It addresses a portion of
https://linear.app/rerun/project/write-tables-to-rerun-cloud-via-the-python-sdk-432c0243348f/overview

### What

With this PR we introduce a new server API to create a table from a
given Schema. The current work will create a Lance based table. This
table is registered with the server and available for immediate writing
using #11565

Additionally this PR adds the integration testing in `re_redap_tests`