Add documentation to dataframe query workflow and fix api rendering for bindings (#11650)
### Related

* Closes RR-2251

### What
* Primary task: first pass through query-data.md to give an overview of
how to start and connect to OSS server and get to a dataframe with
pointers to datafusion for follow on work
* Will fill an issue to make this into a snippet/tutorial to make
testable after we get a python object for the server
* Secondary task: Render docs for DataframeQueryView and lots of other
bindings exposed through catalog
* This scope creeped a lot so I just did the manual class list will file
a follow on ticket for proposal to fix this more globally
1. Add lint codes so I can separate out my ignores for pyclass_eq and
pyclass_module requirements
1. The lint.py changes (which aren't great since I'm not an experience
parser writer but don't seem awful)
3. Add a pyclass module requirement to specify
`rerun_bindings.rerun_bindings` see
https://github.com/rerun-io/rerun/pull/11268 for more context
      1. All the rust file changes 
   4. Fight mkdocs to render bindings properly
      1. The mkdocs related files
2. Adds the ability to generate our docs just off of the stubs without
the shared object being present at all. NOTE: The docs actually look
better this way, but I added some accommodation to make them look less
awful if the shared object is present.

Filed an issue to fix all the other missing classes from our rendered
api docs:
https://linear.app/rerun/issue/RR-2766/cover-remaining-classes-in-python-api-rendered-docs
and an issue with a proposal to avoid this problem moving forward:

https://linear.app/rerun/issue/RR-2765/change-python-public-api-surface-pattern-to-support-easier-docs