Add client object table writing functions (#11657)
### Related

This is a follow on to https://github.com/rerun-io/rerun/pull/11565
This is part of
https://linear.app/rerun/project/write-tables-to-rerun-cloud-via-the-python-sdk-432c0243348f/overview

### What

This PR adds two methods to the client interface in Python:

`write_table` takes in a table name and record batches and writes them
to a table. It contains convenience functions to accept either a record
batch, iterable, iterable of iterable, or a record batch reader.
Additionally you can set it to append or overwrite the table.

`append_to_table` is a convenience function that takes Python objects
and turns them into a row or rows of data. It uses pyarrow for type
casting of the data into arrow arrays. Then it composes a record batch
and writes it to the server by calling `write_table`.