Our IndexValuesLike type is overly strict. Allow np.datetime64 inputs (#11721)
### What
We accept indexes and time specs in a few different places and we seem
to re-implement this handling in a few different ways. This doesn't
resolve that but it does add a test highlighting the lack of ergonomics,
then fixes it here.

A few that I am aware of are:
* Time column:
https://github.com/rerun-io/rerun/blob/main/rerun_py/tests/unit/test_time_column.py
* Dataset entry start/stop:
https://github.com/rerun-io/rerun/blob/7c3d2b05255883ef61d561c7923ea7716c402536/rerun_py/src/catalog/dataset_entry.rs#L229
* This here
* I know Andreas also mentioned a reference implementation for this but
I think that was more internal than sdk python focused

I've contributed to some of the mess and I'll follow a ticket to try and
unify our time handling so we can be as robust and flexible as possible
while still being correct.