Create and write tables in the Rerun server (#11694)
This is the final PR in the group associated with creating and writing
tables in the server. The previous PRs include:

- https://github.com/rerun-io/rerun/pull/11565
- https://github.com/rerun-io/rerun/pull/11640
- https://github.com/rerun-io/rerun/pull/11657

The above mentioned PRs add in the core functionality required for users
to both create new tables and to write to existing tables in the Rerun
server. This PR adds the final step, a usage example. The purpose of
this example is to demonstrate a common work flow in data analysis where
you wish to incrementally evaluate a dataset and write results into a
table.

This PR closes
https://linear.app/rerun/initiative/multistep-robotics-pipelines-f45b92ccebb6/overview

---------

Co-authored-by: ntjohnson1 <24689722+ntjohnson1@users.noreply.github.com>