Fix 3D eye speed-up modifier not working with scroll-to-zoom (#11814)
This is basically caused by egui's scroll behavior that changes when you
press certain modifiers (see
[here](https://github.com/emilk/egui/blob/31d313572b798d1924329cc074f0f22075fca712/crates/egui/src/input_state/mod.rs#L86)).

Pressing shift for horizontal scroll is useful in other views like the
dataframe view, but in the spatial view it conflicts with the speed-up
modifier.

This change ensures that scrolling always zooms the eye.

Fixes: #11813

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>