Make `RecordingStream::disabled()` a `const fn`.
In the implementation, a disabled stream is now represented by
`Weak::new()` instead of `Arc::new(None)`; the `Option` has been removed
as unnecessary.

This will allow users to create `static`s or `const`s which contain
disabled `RecordingStream`s. It also means that creating disabled
streams does not allocate any memory, which may be useful for efficiency
in applications where communicating with Rerun is optional.