Add notebook execution to ci (#11827)
### Related
RR-1880 continue to add better support for our notebook testing

### What
* Adds a slight improvement to the OSS server docs (I was trying to make
progress on this but a notebook tutorial is probably the best direction
for that)
* Updates the nbstripout helper that I added before. Notebooks can bloat
the repo with all the silly metadata changes so make it easy to remove
and make sure it is removed. This runs in ~ 0.5s on my local machine
* Added `notebook` target under `build-examples` this is quite similar
to `rrd` but I don't yet have a clear picture on a full design for how
we want to support notebooks and oss server stuff.
* For now just using nbconvert to execute the notebooks. This is just a
smoke test if anything issues a warning or crashes. We could always
improve the infra further.
* I moved all the notebooks to the new tested notebook directory (we
could probably remove the notebook prefixes from the subdirectories now)
* Only functional changes are removing the native viewer from the
send_tables notebook example and changing over a few deprecated args

### TODO
* [x] `@rerun-bot full-check`
* [x] Trigger `Nightly` on this PR

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>