Add example for loading MCAP with custom messages using Lenses (#11789)
Builds on the draft #11635, i.e. you can load MCAP files with custom
Protobuf
(here: Foxglove messages), convert and send them to Rerun via Lenses
API.

This is an example for an in-progress API, so it's set to ignored in the
`manifest.toml`.

See `README.md` for more information & usage.

https://linear.app/rerun/issue/RR-2851

---------

Co-authored-by: Jochen Görtler <me@jgoertler.com>
Co-authored-by: Jochen Görtler <grtlr@users.noreply.github.com>