Add shortcuts to switch between recordings (#11637)
### Related

* part of
https://linear.app/rerun/issue/RR-2362/navigating-through-catalog-via-recordings-side-panel-with-hotkeys

### What

Adds global hotkeys to switch through everything in the recording panel
(except for the homepage and anything without it's own page, e.g. app
ids).

This was surprisingly deep, the implementation feels quite complex but
is the best thing I could come up with. Its the result of the following
things:
- shortcuts should be a `UiCommand` for discoverability (so it can be
found in the command palette)
- recording switching order should be defined in the recording panel,
since it should match the order of the shortcuts

=> We introduce a new `RecordingPanelCommand` that can be send from a
`UiCommand` and executed in the context of the recording panel. The
latter is now a struct to hold the command queue.

Note:
- We only shift through recordings in the current "goup" (aka dataset or
app id)
- This feature overlaps a lot with the new tree navigation (arrow
keys)....
- ...but could be made massively more useful if we attempted to
synchronise the viewer state:
  - https://github.com/rerun-io/rerun/issues/11792

---------

Co-authored-by: Antoine Beyeler <antoine@rerun.io>