Rename `ControlEye` to `EyeController` (#11895)
Improves readability. The new name makes it clearer that it's only an
utility for mutating eye controls, not an eye itself.

Also renames awkward constructs like `eye.get_eye()` to
`eye_controller.get_eye()`.

This is a cleanup in preparation for
https://linear.app/rerun/issue/RR-178, where we noticed that the current
eye code is a bit confusing to read (this might not be the last pure
refactor, but let's go step by step)