Revise Lenses API and definitions to include scatter columns (#11871)
### What

This PR revises the Lenses API and specification by introducing
different types of lenses that are centered around the time columns:

* `OneToOne` lenses have the same amount of output rows as their input
column.
* `Static` lenses don't have any time columns attached to them.
* `OneToMany` lenses produce multiple rows for any given row in their
input column.

These kind of lenses are accessed in the `LensBuilder` via `columns`,
`static_columns`, and `scatter_columns` methods, which create scopes
that can be used to bundle columns together.

Although the concept of chunks is not exposed in the Lenses API, under
the hood all columns (time or component) specified in a scope
(`LensKind`) will end up in the same chunk.

### Context

The new scatter columns operating mode is required for "plural" MCAP
messages that have multiple timestamps defined in a single message. An
example for this is
[`foxglove.FrameTransforms`](https://docs.foxglove.dev/docs/sdk/schemas/frame-transforms).