Add ability to replace/upsert in tables (#11812)
### Related

Closes
https://linear.app/rerun/project/add-ability-to-upsert-data-in-tables-aad5989947ac

### What

This PR follows up on the work from
https://github.com/rerun-io/rerun/pull/11694 and the issues linked
inside it.

With this PR you can update existing tables if they have an index
specified as a primary key. You do this by setting the metadata
`rerun:is_table_index` on the field when creating the table. During
`update` operations we will use these columns as the keys by which we
update the underlying lance table.