Introduce re_table, implementing styling and basic table things (#11893)
### Related

* part of
https://linear.app/rerun/issue/RR-2360/multi-selection-within-a-partition-table

### What

Simplifies the datafusion table code by pulling out the fuctionality
that doesn't need to know about datafusion in a re_table wrapper. It
handles styling, column order / visibility and eventually row selection.
For now I'll leave it in the datafusion table crate, in case we do find
something in re_table that does require tighter integration with
datafusion. But eventually it could become a separate crate, and be used
for the dataframe table as well to add selection / row numbers / table
config.