Allow to import URDF file into current recording
Previous behavior: if you dragged the URDF file into the viewer,
it created its own new recording. Same for "Import into current
recording" menu (note: requires #11941), which is counter-intuitive.

A recording containing only an URDF is pretty useless, and it's
more common that you have some recording and want to add the model
to it.