Fix sharing links to time ranges on duration timelines (#11948)
This was severely broken because of failure to parse URLs with `+` in
them (e.g. `+1.2s`). I've now fixed the problem(s) and added several
tests.

On my way to fixing this I also attacked a bunch of places where we
`debug` logged some very verbose stuff, that is now actually readable
instead.

See commit messages for details.