Telemetry: Allow initializing grpc_on_eos metrics
The server-side grpc_on_eos metric is generated in a callback. As a
result, it doesn't exist until the first api call, making it impossible
to observer the transition from zero to one or to differentiate "no
request" from "this metric doesn't work".

This change adds an option to pass the endpoint names and if done so,
the metric will get initialized.