Telemetry: Allow initializing grpc_on_eos metrics (#11982)
The server-side grpc_on_eos metric is generated in a callback. As a
result, it doesn't exist until the first api call, making it impossible
to observer the transition from zero to one or to differentiate "no
request" from "this metric doesn't work".

This change adds an option to pass the endpoint names and if done so,
the metric will get initialized.

### Related

<!--
Include links to any related issues/PRs in a bulleted list, for example:
* Closes #1234
* Part of #1337
-->

### What

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.

For more details check the PR section on
<https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md>.
-->