Add shortcuts to navigate the timeline (#11933)
### Related

* closes
https://linear.app/rerun/issue/RR-2391/faster-timeline-navigation-with-shortcuts

This adds shortcuts to navigate the timeline.

The playback speed chord is implemented as a UICommand, so that it is
discoverable via the CommandPalette. It's a bit odd since it has a
value, but I think it works quite nicely.

Gaps aren't handled yet since that would require a bigger refactor of
how we handle them. I've opened an issue here:
https://linear.app/rerun/issue/RR-3023/correctly-handle-timeline-gaps

The hack to override eguis built in shortcuts are quite bad, I intend to
fix this with the next egui release.