Fix `ChunkId` clash due to chunk splitting (#12008)
**Review commit-by-commit!**

We recently introduced chunk splitting on-ingest. The code used the very
dangerous `ChunkId::next`, which can easily cause clashes.

The problem is this: we split chunks that are over some byte limit in
even pieces, where "even" is measured in _row count_. This means that
some of the smaller pieces may still be over the size limit. The caller
(`ChunkStore::insert_chunk`) will in these cases recurse, i.e. split the
smaller pieces further.

This case WILL cause `ChunkId` collisions, if the split algo uses
`ChunkId::next`, which it did.

## Background
I wanted to understand the code a bit more, and also make it re-usable
for fixing https://github.com/rerun-io/rerun/issues/11993

## Related
* Code added in https://github.com/rerun-io/rerun/pull/11921
* The refactor unblocks a quick-fix for
https://github.com/rerun-io/rerun/issues/11993
* Related: https://github.com/rerun-io/rerun/issues/11971