Add `re_arrow_combinators::reshape::RowMajorToColumnMajor` (#12031)
Required for example to read matrices from variable sized, row-major
protobuf lists, e.g. the 3x3 or 3x4 camera matrices that are found in
`foxglove.CameraCalibration` schema.

In order to do that, we need both a cast to fixed-size-list and a
reshaping transformation.

---

This is a prerequisite for #12035.

Contributes to:
https://linear.app/rerun/issue/RR-3045
https://linear.app/rerun/issue/RR-2327