Fix transparency in 2D views (#12214)
* Closes https://github.com/rerun-io/rerun/issues/12023
* Closes https://github.com/rerun-io/rerun/issues/12206
* Closes https://github.com/rerun-io/rerun/issues/11156
* Closes RR-3179
* Closes RR-3057
* Bug introduced in https://github.com/rerun-io/rerun/pull/11962

~I don't understand _why_ this PR fixes the problem, but it does.~

----

@Wumpf : reverted the original fix and applied a more forward looking
fix in the re_renderer: we're now doing the same transparency handling
on `re_renderer::Rectangle` (the underlying primitive) that we do on
other primitives like meshes. I previously assumed that this isn't
necessary because we sort the rectangle draw data, was apparently wrong
either way.

The situation is still a bit weird because we still rely on depth based
draw ordering - our draw order gets translated into a depth offset. With
this change, the ordering for transparent objects take care of it. But
drawbacks with this approach remain:
* depth offset is inaccurate and causes glitches especially on web. This
is probably reduced now since we no longer draw transparent rectangles
with depth (which we did before, relying on their correct ordering). See
#4787
* draw order doesn't always interact correctly between opaque and
non-opaque rectangles
* draw order doesn't always interact correctly between rectangles and
other things with draw order

Solution to all these problems is to have a special 2D draw phase in
which we don't rely on the depth buffer at all but exclusively on an
explicit draw order (which the Viewer already has for 2D objects!).
Other related issues: #1025, #11156

---------

Co-authored-by: Gijs de Jong <berichtaangijs@gmail.com>
Co-authored-by: Andreas Reich <r_andreas2@web.de>